requirements.txt
Django>=4.2
pyyaml
django-filter
django-tables2
django-taggit
habanero
markdown

